third_party/WebKit/LayoutTests/editing/inserting/insert-div-010.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>|x</div>',
        '</div>',
    ].join(''),
    'insertParagraph',
    [
        '<div contenteditable>',
            '<div><br></div>',
            '<div>|x</div>',
        '</div>',
    ].join('')),
    'insert BR before a character');
</script>
